requirements.txt
# API
requests==2.31.0
httpx==0.27.0

# Áudio
sounddevice==0.4.6
scipy==1.11.4
numpy==1.26.2
pydub==0.25.1

# Captura de tela
mss==9.0.1
Pillow==10.1.0

# Google Drive
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Configuração
pyyaml==6.0.1
python-dotenv==1.0.0

# Agendamento
schedule==1.2.1

# Windows
pywin32==306

# Utilidades
python-dateutil==2.8.2
pynput==1.7.6
